Setup automated CI releases
Previously, I have been releasing the plugin from my local computer
using a manual process. This was problematic because it was easy to
screw up the release by having a wrong version of Cody cloned locally,
or forgetting to publish an accompanying alpha release for every stable
release. The 3.4.0 had a critical regression that was caused by this
exact root problem.

This PR is the first step towards addressing the problem by adding CI
workflows to publish releases. By publishing the releases in CI, we
document all the necessary steps to publish a release, and we enable
anyone on the team to cut a release at any time.

By default, we only publish alpha releases for now. I want to follow up
with a separate PR that automates the process to "promote" an alpha
release to stable. My goal is that we use a byte-by-byte identical
release to the alpha when promote to stable.