Stop using Gradle cache for sourcegraph+cody clones
Previously, we did shallow git clones of sourcegraph/sourcegraph and
sourcegraph/cody to the Gradle cache. This was problematic because it
was taking a very long time to create the cache in CI, defeating the
point of caching. This PR moves the shallow clones to another directory
so that they are not cached, which hopefully speeds up the caching step
in CI (which is taking 5-7 minutes before this PR)