Fix version numbering script
Previously, the patch number was computed based on the release timestamp
of the last Sourcegraph release. This was problematic because it meant
we reset the patch number for every new patch release of Sourcegraph.
This PR fixes the problem by making the patch number be based on the
release timestamp of the latest MAJOR.MINOR.0 release.