Add more intelligent agent rebuilding
Previously, we didn't rebuild the agent if the binaries were available
on disk. This was problematic because it meant that we don't
automatically rebuild the binaries when bumping up the Cody commit. This
PR fixes the problem by generating a new `cody-revision.txt` file
alongside the agent binaries. This revision file includes the cody
commit hash, and we automatically rebuild the binaries if this commit
hash doesn't match the cody commit hash in the build.