Agent: move from recipes to the new chat UI
Previously, the JetBrains plugin used the `chat-question` recipe to power chat.
This was problematic because this recipe was not used by VS Code meaning
that users had a different experience in JetBrains and VS Code, usually
a worse experience in JetBrains because it didn't keep up with the rapid
context improvements in VS Code.

Now, with this PR, JetBrains uses new `chat/*` and `webview/*` related
endpoints in the agent to power chat. This PR requires the changes from
https://github.com/sourcegraph/cody/pull/2396 to be merged. The new chat
UI in VS Code is entirely designed around webviews that have a low-level
notification-based protocol (no requests). This low-lever protocol is a
bit more complicated to work with but it supports richer functionality,
including triggering indexing of local embeddings, manually adding
individual files to the context, selecting the language model, and more.
This PR only implements the basics: submitting a message, stopping a
reply, and handling rate limit errors.

Importantly, this PR alone won't close the full gap against VS Code for
chat or commands. This is just an important milestone towards that goal.