Enterprise: add ability to disable logging events to dotcom
Previously, the JetBrains plugin always sent a request to
sourcegraph.com to log events. This was problematic for some enterprise
users on airgapped instances where the agent was failing to initialize
because a request to sourcegraph.com was timing out.
This PR adds a VM option
`-Dcody.log-events-to-connected-instance-only=true` that disables this
behavior. This VM options is a short-term solution. The ideal solution
is that the Cody agent automatically turns on this logging mode based on
information from the site instance, which is configured by the site
admin.