Display onboarding tooltip on first autocomplete suggestion
Fixes #1319.

Previously, users didn't get any onboarding instructions when using Cody
for the first time in JetBrains. For example, many users didn't know
they should press TAB to accept an autocomplete suggestion, or
option-bracket to cycle through options. This PR fixes the problem by
using the built-in `GotItTooltip` component, which is designed for this
use-case.