QA agent using new JSON-RPC library
The PR at https://github.com/sourcegraph/cody/pull/3807 switches the JSON-RPC library used by the agent. This should not change anything, but just to be safe, we want to run a QA test on it.

This PR is not intended to be merged. If the QA test passes, we will merge that PR and then point this JetBrains repo at the commit on `main`.