Onboarding tooltip: don't hide the code
Previously, there were cases when the position of the onboarding tooltip
would hide the underlying code. This goes against JB recommendations
https://jetbrains.design/intellij/controls/got_it_tooltip/#location even
if we are using the built-in position helpers. This PR changes the
position dynamically based on

- Whether the project window is open
- Whether the cursor is close to the top

Depending on the different values, we now either display the tooltip to
the left, above, or below.