implemented Cody Ignore for inline edits
The basic approach was to use IntelliJ's MessageBus pub/sub
to be notified whenever there is a policy change. Inline Edits has
slightly different needs; it only needs to check the policy when the
user starts an Edit, and be notified if the policy subsequently
changes while there is an active fixup session. So pubsub was a
pretty good option here. It will also facilitate testing.