Agent: bump commit
This bumps Cody to a commit that includes this PR here https://github.com/sourcegraph/cody/pull/4279

This PR fixes a tricky issue related to how
`vscode.TextEditor.selection` works. Previously, the text editor
selection was a constant, it never updated to reflect the latest
selection of the document. This could have caused buggy behavior in
scenarios like this:

```ts
for (const editor of vscode.window.visibleTextEditors) {
  // editor.selection is correct
  await parseDocument(editor.document)
  // editor.selection is outdated because we received a
  // `textDocument/didChange` in the meantime
}
```