feat(agent): add server-sent models feature
This change adds a new `serverSentModels` feature to the `ConfigFeatures` class, which is used to track whether the server supports sending models directly to the client. The `LlmDropdown` class is updated to subscribe to changes in the `CurrentConfigFeatures` and show/hide the dropdown based on the value of `serverSentModels`. Additionally, the `CodyAgentClientTest` and `CurrentConfigFeaturesTest` are updated to reflect the new feature.