Put in some fixes for the new cody-highlights code path
It was throwing at least two exceptions, one reported by QA and one that I saw myself.

The one I saw was that ProtocolTextDocument must be created on the EDT.

I don't really know how to reproduce or exercise this code path, so we'll have to wait for Rik to look at it, but it should at least not throw errors now.