Agent: add configuration for server-managed global state and multiple webview instances
This change adds support for server-managed global state and multiple webview instances in the Cody agent.

The `globalState` capability is set to `Server-managed`, which allows the server to manage the global state of the Cody agent. This is useful for Chat History. However, I currently have it commented out as we will need to make sure the chat history is migrated to the new global state.

The `view` property of the `WebviewNativeConfigParams` is set to `Multiple`, which enables the support of multiple webview instances that JetBrains has implemented. This is necessary for features that require multiple webview windows, such as the chat functionality.